+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## In progress
 * Fix the algorithm for marking all users as checked/unchecked. Refs UIPFU-115. 
 * Refactor user filtering logic in UserSearchContainer. Fixes UIPFU-107.
+* Improve UserSearchView checkbox logic to avoid displaying "select all" checkbox checked on next page and be able to select all users from multiple pages. Fixes UIPFU-94.
 
 ## [8.0.0] (https://github.com/folio-org/ui-plugin-find-user/tree/v8.0.0) (2025-03-13)
 [Full Changelog](https://github.com/folio-org/ui-plugin-find-user/compare/v7.2.0...v8.0.0)